requirements.txt
python-dotenv
azure-identity
azure-ai-projects
openai
azure-cli
streamlit
whatchdog

